# restaurantServices/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Restaurant services",
    "description": "Restaurant Services API documentation "
  },
  "host": "localhost:8081",
  "basePath": "/",
  "tags": [
    {
      "name": "restaurant"
    }
  ],
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "bearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "bearerFormat": "JWT",
      "in": "header"
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/restaurant/crete-restaurant": {
      "post": {
        "tags": [
          "restaurant"
        ],
        "description": "",
        "parameters": [
          {
            "name": "newRestaurant",
            "in": "body",
            "required": true,
            "type": "object",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "new- wood"
                },
                "phone": {
                  "type": "number",
                  "example": 7889897889
                },
                "location": {
                  "type": "object",
                  "properties": {
                    "lat": {
                      "type": "string",
                      "example": "224342"
                    },
                    "long": {
                      "type": "string",
                      "example": "432323"
                    }
                  }
                },
                "address": {
                  "type": "string",
                  "example": " pune"
                },
                "cuisine": {
                  "type": "array",
                  "example": [
                    "veg"
                  ],
                  "items": {
                    "type": "string"
                  }
                },
                "menu": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "example": "111"
                      },
                      "name": {
                        "type": "string",
                        "example": "malvani fish"
                      },
                      "price": {
                        "type": "string",
                        "example": "500"
                      }
                    }
                  }
                },
                "rating": {
                  "type": "number",
                  "example": 4
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/restaurant/update-restaurant": {
      "put": {
        "tags": [
          "restaurant"
        ],
        "description": "",
        "parameters": [
          {
            "name": "newRestaurant",
            "in": "body",
            "required": true,
            "type": "object",
            "schema": {
              "type": "object",
              "properties": {
                "cuisine": {
                  "type": "array",
                  "example": [
                    "veg"
                  ],
                  "items": {
                    "type": "string"
                  }
                },
                "name": {
                  "type": "string",
                  "example": "new- wood"
                },
                "phone": {
                  "type": "number",
                  "example": 7889897889
                },
                "location": {
                  "type": "object",
                  "properties": {
                    "lat": {
                      "type": "string",
                      "example": "224342"
                    },
                    "long": {
                      "type": "string",
                      "example": "432323"
                    }
                  }
                },
                "address": {
                  "type": "string",
                  "example": " pune"
                },
                "rating": {
                  "type": "number",
                  "example": 4
                },
                "createdAt": {
                  "type": "string",
                  "example": "2021-08-04T07:58:38.903Z"
                },
                "updatedAt": {
                  "type": "string",
                  "example": "2021-08-04T07:58:38.903Z"
                },
                "id": {
                  "type": "string",
                  "example": "610a48ae5bb80d1128cd9aed"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/restaurant/find-all-restaurant": {
      "get": {
        "tags": [
          "restaurant"
        ],
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/restaurant/search": {
      "post": {
        "tags": [
          "restaurant"
        ],
        "description": "",
        "parameters": [
          {
            "name": "newRestaurant",
            "in": "body",
            "required": true,
            "type": "object",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "example": "610a638ed1b5311b5cd498eb"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/restaurant/delete": {
      "delete": {
        "tags": [
          "restaurant"
        ],
        "description": "",
        "parameters": [
          {
            "name": "newRestaurant",
            "in": "body",
            "required": true,
            "type": "object",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "example": "6108f2a804457a3f14d57c6e"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  },
  "definitions": {},
  "security": [
    {
      "bearerAuth": []
    }
  ]
}